Implement Application Factory pattern
