+++ AIPscan/Aggregator/tasks.py
@@ -6,8 +6,9 @@ import requests
 
 from celery.utils.log import get_task_logger
 
-from AIPscan import celery
+# from AIPscan.extensions import celery
 from AIPscan import db
+from AIPscan.extensions import celery
 from AIPscan.models import (
     FetchJob,
     # Custom celery Models.
