+++ AIPscan/worker.py
@@ -1,0 +1,5 @@
+from AIPscan.application import create_app
+from AIPscan.celery import configure_celery
+
+app = create_app()
+celery = configure_celery(app)
