+++ create_aipscan.db.py
@@ -4,5 +4,8 @@
 # ./create_aipscan.db.py
 
 from AIPscan import db
+from AIPscan.application import create_app
 
+app = create_app()
+app.app_context().push()
 db.create_all()
