+++ run.py
@@ -1,4 +1,6 @@
-from AIPscan import app
+from AIPscan.application import create_app
+
 
 if __name__ == "__main__":
+    app = create_app()
     app.run(debug=True, host="0.0.0.0")
